Moving constants file to the root
