+++ amundsen_application/static/js/constants.ts
