+++ amundsen_application/static/js/pages/DashboardPage/index.spec.tsx
@@ -13,8 +13,8 @@ import BookmarkIcon from 'components/common/Bookmark/BookmarkIcon';
 import ResourceList from 'components/common/ResourceList';
 import TabsComponent from 'components/common/TabsComponent';
 import { dashboardMetadata } from 'fixtures/metadata/dashboard';
-import { NO_TIMESTAMP_TEXT } from 'components/constants';
 import { ResourceType } from 'interfaces';
+import { NO_TIMESTAMP_TEXT } from '../../constants';
 import ChartList from './ChartList';
 import DashboardOwnerEditor from './DashboardOwnerEditor';
 import ImagePreview from './ImagePreview';
